[JT-78] refactor: 로그인 컨트롤러 변경
- 기존 AuthController에서 MemberController로 로그인 위치 변경

- 불필요한 Smtp 파일 삭제
