+++ module-application/app-api/src/main/java/shop/jtoon/member/presentation/MemberController.java
@@ -9,10 +9,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import shop.jtoon.member.application.MemberApplicationService;
 import shop.jtoon.member.request.SignUpReq;
+import shop.jtoon.security.request.LoginReq;
 
 @RestController
 @RequiredArgsConstructor
@@ -28,7 +30,13 @@ public class MemberController {
 	}
 
 	@GetMapping("/email-authorization")
-	public String authenticateEmail(@RequestParam(value = "email") String email) {
-		return memberApplicationService.sendEmailAuthentication(email);
+	@ResponseStatus(HttpStatus.CREATED)
+	public void authenticateEmail(@RequestParam(value = "email") String email) {
+		memberApplicationService.sendEmailAuthentication(email);
+	}
+
+	@PostMapping("/local-login")
+	public void login(@RequestBody @Valid LoginReq loginReq, HttpServletResponse response) {
+		memberApplicationService.loginMember(loginReq, response);
 	}
 }
